Check-in scene file parser
This partially implements a parser for a custom scene file format. The format resembles JSON with some tweaks to make it more readable.

I initially tried to use serde and serde_json to handle the parsing, but it seemed that we would have to do most of the parsing work on the JSON anyway, so I decided to write it from scratch instead.

I've ended up with a design that relies heavily on the TryFrom trait to convert from "raw" values to various expected types like f64 and Color. This is a a bit magical, but it makes the higher-level parsing code more concise and easy to write. See the TryFrom implementation for Material as an example.

+++ src/scene.rs
@@ -6,7 +6,7 @@ use crate::{
 };
 
 pub struct Scene {
-    pub max_depth: u32,
+    pub max_depth: usize,
     pub film_width: usize,
     pub film_height: usize,
     pub camera: Box<dyn Camera>,
@@ -15,7 +15,7 @@ pub struct Scene {
 
 impl Scene {
     pub fn new(
-        max_depth: u32,
+        max_depth: usize,
         film_width: usize,
         film_height: usize,
         camera: Box<dyn Camera>,
